pyproject.toml
[project]
name = "mcp-gateway"
version = "0.1.0"
description = "Production-grade MCP Gateway with OAuth, RBAC, human approval, and append-only audit"
requires-python = ">=3.13"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "Roman Serbin" }]
dependencies = [
    "fastmcp>=2.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "sqlalchemy[asyncio]>=2.0.35",
    "asyncpg>=0.30.0",
    "alembic>=1.13.0",
    "authlib>=1.3.0",
    "pyjwt[crypto]>=2.9.0",
    "cryptography>=43.0.0",
    "httpx>=0.27.0",
    "tenacity>=9.0.0",
    "python-telegram-bot>=21.6",
    "jinja2>=3.1.0",
    "structlog>=24.4.0",
    "prometheus-client>=0.21.0",
    "opentelemetry-api>=1.27.0",
    "opentelemetry-sdk>=1.27.0",
    "opentelemetry-exporter-otlp>=1.27.0",
    "opentelemetry-instrumentation-fastapi>=0.48b0",
    "opentelemetry-instrumentation-asyncpg>=0.48b0",
    "opentelemetry-instrumentation-httpx>=0.48b0",
    "pyyaml>=6.0",
    "python-multipart>=0.0.10",
    "websockets>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "testcontainers[postgres]>=4.8.0",
    "respx>=0.21.1",
    "mutmut>=3.0.0",
    "locust>=2.31.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
    "black>=24.8.0",
    "bandit>=1.7.0",
    "pip-audit>=2.7.0",
    "pre-commit>=3.8.0",
    "types-pyyaml",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["gateway"]

[tool.ruff]
line-length = 100
target-version = "py313"
exclude = ["alembic/versions"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "S", "A", "C4", "T20", "RET", "SIM", "ARG"]
# N818: legacy Token* exception names without Error suffix (renaming breaks downstream).
# ARG001: framework callback signatures (telegram, fastapi) require unused params.
# E402: conditional imports inside functions (telegram, optional deps).
ignore = ["S101", "B008", "S104", "N818", "ARG001", "E402"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S", "ARG", "T20"]
"mocks/**" = ["S104", "T20"]
"gateway/cli.py" = ["T20"]

[tool.mypy]
python_version = "3.13"
# Phased typing: --strict is a future goal. Currently catches signature issues
# without forcing a one-shot annotation pass on every dict literal.
plugins = ["pydantic.mypy"]
exclude = ["alembic/versions"]
warn_unused_ignores = true
warn_redundant_casts = true
check_untyped_defs = true
no_implicit_optional = true
disallow_untyped_calls = false
disallow_untyped_defs = false
disallow_incomplete_defs = false

[[tool.mypy.overrides]]
module = ["testcontainers.*", "fastmcp.*", "telegram.*", "respx.*", "locust.*", "asyncpg.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"
testpaths = ["tests"]
markers = [
    "unit: fast tests with no I/O",
    "integration: tests requiring containers",
    "e2e: full stack tests via docker compose",
    "security: security-focused tests",
]

[tool.coverage.run]
source = ["gateway"]
branch = true

[tool.coverage.report]
# Unit-only floor. Integration tests cover web/middleware/server/audit/approval modules
# but their coverage isn't merged into the unit job in CI yet (TODO: --cov-append).
fail_under = 35
show_missing = true
